+++ types.go
@@ -23,7 +23,7 @@ type DBInterface interface {
 
 	NewThread(name, domain, owner, iteration, state string, percentile float64, cost int) (int64, error)
 	LinkThreads(parent, child int64, iter string, ord int, domain string) error
-	LinkThreadsStakeholder(parent, child int, stakeholder, domain string) error
+	LinkThreadsStakeholder(parent, child int64, stakeholder, domain string) error
 	GetThreadOrderBefore(parent int64, iter string, order int) (int, error)
 
 	GetThreadrel(id int64, stakeholder string) (*taps.Threadrel, error)
